Bug 30970: Add checkbox column to holdst and holdso
This patch adds the missing checkbox column to the columns_settings
configuration file.

The field was added unconditionally to the tables in bug 23678 and then
consequently fixed in master as part of bug 27868 which being an
ehancement wont be backported.  This patch simply lifts the relevant
part of the fix from bug 27868 to apply just to the 21.11.x branch.

Signed-off-by: Lucy Harrison <l.m.harrison@lboro.ac.uk>
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Victor Grousset/tuxayo <victor@tuxayo.net>
Signed-off-by: Arthur Suzuki <arthur.suzuki@biblibre.com>